add message parameter to @deprecated
