+++ docs/pedantic/index.html
@@ -64,10 +64,6 @@ el.replaceWith(d);
 <dd>
 <div class="desc"></div>
 </dd>
-<dt><code class="name"><a title="pedantic.helper_methods" href="helper_methods.html">pedantic.helper_methods</a></code></dt>
-<dd>
-<div class="desc"></div>
-</dd>
 <dt><code class="name"><a title="pedantic.mixins" href="mixins/index.html">pedantic.mixins</a></code></dt>
 <dd>
 <div class="desc"></div>
@@ -102,7 +98,6 @@ el.replaceWith(d);
 <li><code><a title="pedantic.examples" href="examples/index.html">pedantic.examples</a></code></li>
 <li><code><a title="pedantic.exceptions" href="exceptions.html">pedantic.exceptions</a></code></li>
 <li><code><a title="pedantic.get_context" href="get_context.html">pedantic.get_context</a></code></li>
-<li><code><a title="pedantic.helper_methods" href="helper_methods.html">pedantic.helper_methods</a></code></li>
 <li><code><a title="pedantic.mixins" href="mixins/index.html">pedantic.mixins</a></code></li>
 <li><code><a title="pedantic.models" href="models/index.html">pedantic.models</a></code></li>
 <li><code><a title="pedantic.type_checking_logic" href="type_checking_logic/index.html">pedantic.type_checking_logic</a></code></li>
