+++ pedantic/decorators/cls_deco_frozen_dataclass.py
@@ -142,8 +142,3 @@ def frozen_dataclass(
         return decorator
 
     return decorator(cls_=cls)
-
-
-if __name__ == '__main__':
-    import doctest
-    doctest.testmod(verbose=False, optionflags=doctest.ELLIPSIS)
