+++ pedantic/decorators/fn_deco_deprecated.py
@@ -1,29 +1,40 @@
+import warnings
 from functools import wraps
-from typing import Any
+from typing import Any, Callable, overload
 
 from pedantic.constants import F, ReturnType
-from pedantic.helper_methods import _raise_warning
 
+@overload
+def deprecated(func: F) -> F: ...
 
-def deprecated(func: F) -> F:
+@overload
+def deprecated(*, message: str = '') -> Callable[[F], F]: ...
+
+def deprecated(func: F | None = None, message: str = '') -> F | Callable[[F], F]:
     """
         Use this decorator to mark a function as deprecated. It will raise a warning when the function is called.
+        You can specify an optional reason or message to display with the warning.
 
         Example:
-
         >>> @deprecated
         ... def my_function(a, b, c):
         ...     pass
-        >>> my_function(5, 4, 3)
+        >>> my_function(5, 4, 3)  # doctest: +SKIP
+        >>> @deprecated(message='Will be removed soon. Please use my_function_new_instead.')
+        ... def my_function(a, b, c):
+        ...     pass
+        >>> my_function(5, 4, 3)  # doctest: +SKIP
     """
 
-    @wraps(func)
-    def wrapper(*args: Any, **kwargs: Any) -> ReturnType:
-        _raise_warning(msg=f'Call to deprecated function {func.__qualname__}.', category=DeprecationWarning)
-        return func(*args, **kwargs)
-    return wrapper
+    def decorator(fun: F) -> F:
+        @wraps(fun)
+        def wrapper(*args: Any, **kwargs: Any) -> ReturnType:
+            msg = f'Call to deprecated function {fun.__qualname__}.'
 
+            if message:
+                msg += f'\nReason: {message}'
 
-if __name__ == "__main__":
-    import doctest
-    doctest.testmod(verbose=False, optionflags=doctest.ELLIPSIS)
+            warnings.warn(message=msg, category=DeprecationWarning, stacklevel=2)
+            return fun(*args, **kwargs)
+        return wrapper
+    return decorator if func is None else decorator(func)
