+++ pedantic/decorators/fn_deco_does_same_as_function.py
@@ -52,8 +52,3 @@ def does_same_as_function(other_func: F) -> F:
             return wrapper
 
     return decorator
-
-
-if __name__ == "__main__":
-    import doctest
-    doctest.testmod(verbose=False, optionflags=doctest.ELLIPSIS)
