+++ pedantic/decorators/fn_deco_rename_kwargs.py
@@ -45,8 +45,3 @@ def rename_kwargs(*params: Rename) -> Callable[[F], F]:
             return func(*args, **result_kwargs)
         return wrapper
     return decorator
-
-
-if __name__ == "__main__":
-    import doctest
-    doctest.testmod(verbose=False, optionflags=doctest.ELLIPSIS)
