+++ pedantic/decorators/fn_deco_trace.py
@@ -39,8 +39,3 @@ def trace(func: F) -> F:
         return async_wrapper
     else:
         return wrapper
-
-
-if __name__ == "__main__":
-    import doctest
-    doctest.testmod(verbose=False, optionflags=doctest.ELLIPSIS)
