+++ pedantic/decorators/fn_deco_trace_if_returns.py
@@ -46,8 +46,3 @@ def trace_if_returns(return_value: ReturnType) -> F:
         else:
             return wrapper
     return decorator
-
-
-if __name__ == "__main__":
-    import doctest
-    doctest.testmod(verbose=False, optionflags=doctest.ELLIPSIS)
