+++ tests/decorators/test_small_method_decorators.py
@@ -1,98 +1,11 @@
 import asyncio
 import warnings
-from abc import abstractmethod
 
 import pytest
 
-from pedantic import overrides, timer, count_calls, trace, trace_if_returns, does_same_as_function, deprecated, \
+from pedantic import timer, count_calls, trace, trace_if_returns, does_same_as_function, deprecated, \
     unimplemented, mock, require_kwargs
-from pedantic.exceptions import NotImplementedException, PedanticOverrideException, PedanticCallWithArgsException
-
-
-def test_overrides_parent_has_no_such_method():
-    class MyClassA:
-        pass
-
-    with pytest.raises(expected_exception=PedanticOverrideException):
-        class MyClassB(MyClassA):
-            @overrides(MyClassA)
-            def operation(self): pass
-
-
-def test_overrides_all_good():
-    class MyClassA:
-        def operation(self): pass
-
-    class MyClassB(MyClassA):
-        @overrides(MyClassA)
-        def operation(self):
-            return 42
-
-    b = MyClassB()
-    b.operation()
-
-
-def test_overrides_static_method():
-    class MyClassA:
-        @staticmethod
-        def operation(): pass
-
-    class MyClassB(MyClassA):
-        @staticmethod
-        @overrides(MyClassA)
-        def operation():
-            return 42
-
-    b = MyClassB()
-    assert b.operation() ==  42
-    assert MyClassB.operation() == 42
-
-
-def test_overrides_below_property():
-    class MyClassA:
-        @property
-        @abstractmethod
-        def operation(self): pass
-
-    class MyClassB(MyClassA):
-        @property
-        @overrides(MyClassA)   # Note: it does not work the other way around
-        def operation(self):
-            return 43
-
-    b = MyClassB()
-    assert b.operation == 43
-
-
-def test_overrides_function():
-    class MyClassA:
-        pass
-
-    with pytest.raises(expected_exception=PedanticOverrideException):
-        @overrides(MyClassA)
-        def operation(): return 42
-
-
-def test_deprecated_1():
-    @deprecated
-    def old_method(i: int) -> str: return str(i)
-
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter("always")
-        old_method(42)
-        assert len(w) == 1
-        assert issubclass(w[-1].category, DeprecationWarning)
-        assert "deprecated" in str(w[-1].message)
-
-
-def test_deprecated_2():
-    def old_method(i: int) -> str:
-        return str(i)
-
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter("always")
-        old_method(42)
-        assert not len(w) == 1
+from pedantic.exceptions import NotImplementedException, PedanticCallWithArgsException
 
 
 def test_unimplemented():
@@ -163,32 +76,6 @@ def test_does_same_as_function_wrong():
 
 
 @pytest.mark.asyncio
-async def test_overrides_async_instance_method() -> None:
-    class MyClassA:
-        async def operation(self): pass
-
-    class MyClassB(MyClassA):
-        @overrides(MyClassA)
-        async def operation(self):
-            await asyncio.sleep(0)
-            return 42
-
-    b = MyClassB()
-    await b.operation()
-
-
-@pytest.mark.asyncio
-async def test_overrides_parent_has_no_such_method_async():
-    class MyClassA:
-        pass
-
-    with pytest.raises(expected_exception=PedanticOverrideException):
-        class MyClassB(MyClassA):
-            @overrides(MyClassA)
-            async def operation(self): return 42
-
-
-@pytest.mark.asyncio
 async def test_count_calls_async():
     @count_calls
     async def operation(i: int) -> str:
@@ -228,20 +115,6 @@ async def test_timer_async():
         return str(i)
 
     await operation(42)
-
-
-@pytest.mark.asyncio
-async def test_deprecated_async():
-    @deprecated
-    async def old_method(i: int) -> str:
-        return str(i)
-
-    with warnings.catch_warnings(record=True) as w:
-        warnings.simplefilter("always")
-        await old_method(42)
-        assert len(w) == 1
-        assert issubclass(w[-1].category, DeprecationWarning)
-        assert "deprecated" in str(w[-1].message)
 
 
 @pytest.mark.asyncio
